requirements.txt
# Core dependencies
psycopg2-binary==2.9.9
kafka-python==2.0.2
pandas==2.1.4
numpy==1.26.2
faker==20.1.0
python-dotenv==1.0.0

# AWS
boto3==1.34.7
awscli==1.32.7

# Development
pytest==7.4.3
black==23.12.0
flake8==6.1.0
isort==5.13.2

# Monitoring
prometheus-client==0.19.0

psycopg2-binary==2.9.9
python-dotenv==1.0.0
tqdm